+++ NSArray_DeepMutableCopy.m
@@ -13,7 +13,7 @@
 - (NSMutableArray *)deepMutableCopy
 {
     NSMutableArray *newArray;
-    unsigned int index, count;
+    NSUInteger index, count;
 	
     count = [self count];
     newArray = [[NSMutableArray alloc] initWithCapacity:count];
